Fix compilation without default features
